WebProject/requirements.txt

alabaster==0.7.12
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.6.0
Babel==2.9.1
beautifulsoup4==4.10.0
bootstrap4==0.1.0
cffi==1.15.0
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.7
colorama==0.4.4
coverage==6.1.1
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.45.0
django-background-tasks==1.2.5
django-bootstrap4==21.1
django-compat==1.0.15
django-cors-headers==3.7.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.1
django-discover-runner==1.0
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22
django-entangled==0.4
django-filter==21.1
django-formset==0.1
django-froala-editor==3.2.2
django-ipware==3.0.2
django-nine==0.2.4
django-redis==4.12.1
django-rest-elasticsearch==0.4.2
django-role-permissions==3.1.1
django-webpack-loader==0.7.0
djangorestframework==3.12.4
docutils==0.17.1
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
Faker==8.1.0
flake8==4.0.1
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
ipaddress==1.0.23
Jinja2==3.0.2
lupyne==2.5
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.21.4
oauthlib==3.1.1
packaging==21.2
pandas==1.3.4
pluggy==1.0.0
py==1.10.0
pycodestyle==2.8.0
pycparser==2.21
pyee==8.2.2
pyflakes==2.4.0
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.4.0
pytest-mock==3.6.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
text-unidecode==1.3
tika==1.24
toml==0.10.2
tomli==1.2.2
typing-extensions==3.7.4.3
tzdata==2021.5
urllib3==1.26.4
websockets==10.0